+++ src/maicos/lib/math.py
@@ -600,7 +600,7 @@ def symmetrize(
     return out
 
 
-def compute_form_factor(q: float, atom_type: str) -> float:
+def compute_form_factor(q: float, element: str) -> float:
     r"""Calculate the form factor :math:`f(q)`.
 
     :math:`f(q)` is expressed in terms of the scattering vector as
@@ -618,24 +618,16 @@ def compute_form_factor(q: float, atom_type: str) -> float:
     ----------
     q : float
         The magnitude of the scattering vector in reciprocal angstroms (1/Å).
-    atom_type : str
-        The type of the atom for which the form factor is calculated. The ``atom_type``
-        is attempted to be converted into an element using
-        :obj:`maicos.lib.tables.atomtypes`. If no suitable element is found, it is taken
-        as is.
+    element : str
+        The element for which the form factor is calculated.
 
     Returns
     -------
     float
-        The calculated form factor for the specified atom type and q in units of
+        The calculated form factor for the specified element and q in units of
         electrons.
 
     """
-    if atom_type in tables.atomtypes:
-        element = tables.atomtypes[atom_type]
-    else:
-        element = atom_type
-
     if element == "CH1":
         form_factor = compute_form_factor(q, "C") + compute_form_factor(q, "H")
     elif element == "CH2":