+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ Unreleased
 ----------
 Philip Loche, Henrik Stooß
 
+- Base Cromer-Mann form factor on ``atomgroup.elements`` (#492)
 - Add electron density weights for density modules (#492)
 - Use gallery view for examples (#491)
 - Create logic for stable and latest documentation versions (#489)